requirements.txt
kafka-python
pyspark
streamlit
pandas
plotly
python-dotenv
requests